repr/adapter: move CoordTimestamp to repr + rename to TimestampManipulation
