+++ src/adapter/src/session.rs
@@ -21,7 +21,7 @@ use tokio::sync::mpsc::UnboundedReceiver;
 use tokio::sync::OwnedMutexGuard;
 
 use mz_pgrepr::Format;
-use mz_repr::{Datum, Diff, GlobalId, Row, ScalarType};
+use mz_repr::{Datum, Diff, GlobalId, Row, ScalarType, TimestampManipulation};
 use mz_sql::ast::{Raw, Statement, TransactionAccessMode};
 use mz_sql::plan::{Params, PlanContext, StatementDesc};
 use mz_sql_parser::ast::TransactionIsolationLevel;
@@ -29,7 +29,6 @@ use mz_sql_parser::ast::TransactionIsolationLevel;
 use crate::catalog::SYSTEM_USER;
 use crate::client::ConnectionId;
 use crate::coord::peek::PeekResponseUnary;
-use crate::coord::CoordTimestamp;
 use crate::error::AdapterError;
 use crate::session::vars::IsolationLevel;
 use crate::util::ComputeSinkId;
@@ -56,7 +55,7 @@ pub struct Session<T = mz_repr::Timestamp> {
     drop_sinks: Vec<ComputeSinkId>,
 }
 
-impl<T: CoordTimestamp> Session<T> {
+impl<T: TimestampManipulation> Session<T> {
     /// Creates a new session for the specified connection ID.
     pub fn new(conn_id: ConnectionId, user: String) -> Session<T> {
         assert_ne!(conn_id, DUMMY_CONNECTION_ID);
